Fix DateInput so that formik state is updating onblur
