+++ packages/ui/src/DateInput/DateInput.tsx
@@ -2,13 +2,14 @@ import React, { useState, useEffect } from "react";
 
 import { useTranslation, DateFormat } from "@eisbuk/translations";
 
-import { isISODay } from "../utils/date";
+import { dateToISO, isoToDate } from "../utils/date";
 import TextInput, { TextInputProps } from "../TextInput";
 
 const DateInput: React.FC<TextInputProps> = ({ formikField, ...props }) => {
-  const { field } = formikField;
+  const { field, form } = formikField;
 
-  const { value: inputValue, onChange } = field;
+  const { value: inputValue } = field;
+  const { setFieldValue } = form;
 
   const { t } = useTranslation();
   const [value, setValue] = useState("");
@@ -19,10 +20,6 @@ const DateInput: React.FC<TextInputProps> = ({ formikField, ...props }) => {
     }
   }, [inputValue]);
 
-  useEffect(() => {
-    onChange(value);
-  }, [value]);
-
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setValue(e.target.value);
   };
@@ -30,6 +27,7 @@ const DateInput: React.FC<TextInputProps> = ({ formikField, ...props }) => {
   const handleBlur = (e: React.FocusEvent<HTMLInputElement>) => {
     const isoDate = dateToISO(e.target.value);
     setValue(isoDate);
+    setFieldValue(field.name, isoDate, true);
   };
 
   const _field = {
@@ -37,8 +35,6 @@ const DateInput: React.FC<TextInputProps> = ({ formikField, ...props }) => {
     field: {
       ...field,
       value: isoToDate(value),
-      onBlur: handleBlur,
-      onChange: handleChange,
     },
   };
 
@@ -47,41 +43,10 @@ const DateInput: React.FC<TextInputProps> = ({ formikField, ...props }) => {
       {...props}
       formikField={_field}
       placeholder={t(DateFormat.Placeholder)}
+      onChange={handleChange}
+      onBlur={handleBlur}
     />
   );
 };
-
-/**
- * A helper function that parses ISO string and returns a string in european date format
- * (`dd/mm/yyyy`)
- *
- * _if passed `input` is not a valid ISO string, it returns the `input` as is_
- * @param input ISO string to convert
- * @returns string
- */
-const isoToDate = (input: string): string => {
-  const [year, month, day] = input.split("-");
-  return isISODay(input) ? `${day}/${month}/${year}` : input;
-};
-
-/**
- * A helper function that parses a valid european date separated by one of (. , / , -)
- * into an ISOString
- *
- * _if passed `input` is not a valid ISO string, it returns the `input` as is_
- * @param input user date input
- * @returns string
- */
-const dateToISO = (input: string): string => {
-  const [day, month, year] = input.split(/[-/.]/).map(twoDigits);
-  const isoString = `${year}-${month}-${day}`;
-  return isISODay(isoString) ? isoString : input;
-};
-
-const twoDigits = (value: string) =>
-  Number(value).toLocaleString("en-US", {
-    minimumIntegerDigits: 2,
-    useGrouping: false,
-  });
 
 export default DateInput;
